+++ src/http/mod.rs
@@ -1,5 +1,7 @@
 //! A helper module that supports HTTP
 
+#[cfg(feature = "altair")]
+mod altair_source;
 #[cfg(feature = "graphiql")]
 mod graphiql_plugin;
 #[cfg(feature = "graphiql")]
@@ -14,6 +16,8 @@ mod websocket;
 
 use std::io::ErrorKind;
 
+#[cfg(feature = "altair")]
+pub use altair_source::*;
 use futures_util::io::{AsyncRead, AsyncReadExt};
 #[cfg(feature = "graphiql")]
 pub use graphiql_plugin::{graphiql_plugin_explorer, GraphiQLPlugin};