+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [7.0.6] 2024-06-08
+
+- add license files to each project [#1523](https://github.com/async-graphql/async-graphql/issues/1523)
+- Improve alignment of directive behavior with GraphQL spec [#1524](https://github.com/async-graphql/async-graphql/pull/1524)
+- dynamic schema: pass default vals to ResolverContext [#1527](https://github.com/async-graphql/async-graphql/pull/1527)
+- Add [altair](https://github.com/altair-graphql/altair) source [#1530](https://github.com/async-graphql/async-graphql/pull/1530)
+- feat: Add support for using `Interface` and `OneofObject` on the same struct [#1534](https://github.com/async-graphql/async-graphql/pull/1534)
+
 # [7.0.5] 2024-05-09
 
 - Fix compiler and clippy warnings [#1501](https://github.com/async-graphql/async-graphql/pull/1501)